+++ Assets/GlitchEscape/Scripts/Effects/EffectList.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7b77152230eb45c19e1a66c66634ddd0
+timeCreated: 1587265785
