+++ Assets/GlitchEscape/Scripts/Effects/EffectState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d3d9258554614e12b826fd4dd542510b
+timeCreated: 1587265736
